+++ tests/Fixtures/ConfigMode.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Infocyph\ArrayKit\Tests\Fixtures;
+
+enum ConfigMode: string
+{
+    case Local = 'local';
+    case Prod = 'prod';
+}
